Fix logic for writing playbook extra vars
Previously, the playbook-specific extra variables weren't included in
the extra vars file for a given playbook. This change fixes that bug,
and also updates the path to the extra vars file to by dynamically
generated on a per run basis. This will make the code more robust by
allowing multiple instances of cvengine to be run on the same host at a
time, since there won't be collissions on the file name.