Fix running cvengine with random UID
When running the container in Openshift, a random UID is assigned to the
running container. This causes several issues, most notably an inability
to write to $HOME (required by Ansible), and an inability to SSH to a
remote host, which is critical to Ansible running properly (the ssh
command requires a username be associated with the current UID).

This fix is in line with the best practices that I could find for
resolving the issue. Much of the code was copied out of the sclorg
python s2i image. Specifically, the problem is solved by:

1) Changing group ownership of $HOME to root and then making everything
   under $HOME group writeable. This works because, when Openshift
   assigns a random user, that user is in the root group.
2) Enabling the nss_wrapper package. This package wraps the system
   /etc/passwd file, allowing for overriding its contents. The package
   is first installed (after installing epel, where it's installed
   from). In order for nss_wrapper to take effct, a new passwd file
   needs to be created at container runtime. To facilitate this, I
   added a new script that sets up the passwd file and then calls the
   cvengine script. This script is the new default action when running
   the container without overriding cmd