Add a dockerfile to run a full atomic validation
This is the start of something that will probably be used to execute
container validation in openshift as part of the ci pipeline. It deploys
a container that downloads a fedora atomic ISO, deploys it as a VM in
the container, then runs cvengine against it for a target image.

Longer term this should be modified to facilitate specifying the
image_url, cvdata_url, and artifacts directory at container runtime, but
for testing/demo purposes these are hardcoded for now.

NOTE: Much of this code has been copied out of the ci-pipeline repo.
I've yet to determine exactly how cvengine falls into that Jenkins CI
pipeline, so for now I'm putting all of this here. I expect that at some
point, some of this new code will be integrated into the ci-pipeline
repo, but for now I'm just getting things in version control