Add PreConfiguredEnvironment handler
The preconfigured environment is for when the container platform has
already been deployed and we just need to know the connection
information. This is how all container validations run today, so it
needs to be the first (and default) environment type. This implements
this envionment handler and updates the platform handlers to make use of
environments.